Edit line syntax functions
